+++ tests/test_string.zig
@@ -596,6 +596,54 @@ test "comptime parse with comptime format string, am/pm and 12-hour input" {
     }
 }
 
+test "parse with abbreviated month and day name, locale-specific" {
+    if (!locale_ok()) return error.SkipZigTest;
+
+    const cases = [_]TestCase{
+        .{
+            .string = "Thu 01 Jan 1970, 12 am",
+            .dt = try Datetime.fromFields(.{ .year = 1970 }),
+        },
+        .{
+            .string = "Thu 18 Feb 2021, 11 AM",
+            .dt = try Datetime.fromFields(.{ .year = 2021, .month = 2, .day = 18, .hour = 11 }),
+        },
+        .{
+            .string = "Fri 31 Dec 2021, 5 pm",
+            .dt = try Datetime.fromFields(.{ .year = 2021, .month = 12, .day = 31, .hour = 17 }),
+        },
+    };
+
+    inline for (cases) |case| {
+        const dt = try Datetime.fromString(case.string, "%a %d %b %Y, %I %p");
+        try testing.expectEqual(case.dt, dt);
+    }
+}
+
+test "parse with month name and day, locale-specific" {
+    if (!locale_ok()) return error.SkipZigTest;
+
+    const cases = [_]TestCase{
+        .{
+            .string = "Thursday 01 January 1970, 12 am",
+            .dt = try Datetime.fromFields(.{ .year = 1970 }),
+        },
+        .{
+            .string = "Thursday 18 February 2021, 11 AM",
+            .dt = try Datetime.fromFields(.{ .year = 2021, .month = 2, .day = 18, .hour = 11 }),
+        },
+        .{
+            .string = "Friday 31 December 2021, 5 pm",
+            .dt = try Datetime.fromFields(.{ .year = 2021, .month = 12, .day = 31, .hour = 17 }),
+        },
+    };
+
+    inline for (cases) |case| {
+        const dt = try Datetime.fromString(case.string, "%A %d %B %Y, %I %p");
+        try testing.expectEqual(case.dt, dt);
+    }
+}
+
 test "parse %I and am/pm errors" {
     var err = Datetime.fromString("19 am", "%I %p"); // invalid hour
     try testing.expectError(error.InvalidFormat, err);